[DbService] Partialize
