+++ TheGodfather/Services/DatabaseService/GuildConfigDatabaseService.cs
@@ -1,0 +1,103 @@
+﻿#region USING_DIRECTIVES
+using System;
+using System.Threading.Tasks;
+
+using Npgsql;
+using NpgsqlTypes;
+#endregion
+
+namespace TheGodfather.Services
+{
+    public partial class DatabaseService
+    {
+        public async Task<ulong> GetGuildWelcomeChannelIdAsync(ulong gid)
+        {
+            await _sem.WaitAsync();
+
+            ulong cid = 0;
+
+            using (var con = new NpgsqlConnection(_connectionString))
+            using (var cmd = con.CreateCommand()) {
+                await con.OpenAsync().ConfigureAwait(false);
+
+                cmd.CommandText = "SELECT welcome_cid FROM gf.guild_cfg WHERE gid = @gid LIMIT 1;";
+                cmd.Parameters.AddWithValue("gid", NpgsqlDbType.Bigint, gid);
+
+                var res = await cmd.ExecuteScalarAsync().ConfigureAwait(false);
+                if (res != null && !(res is DBNull))
+                    cid = (ulong)(long)res;
+            }
+
+            _sem.Release();
+            return cid;
+        }
+
+        public async Task<ulong> GetGuildLeaveChannelIdAsync(ulong gid)
+        {
+            await _sem.WaitAsync();
+
+            ulong cid = 0;
+
+            using (var con = new NpgsqlConnection(_connectionString))
+            using (var cmd = con.CreateCommand()) {
+                await con.OpenAsync().ConfigureAwait(false);
+
+                cmd.CommandText = "SELECT leave_cid FROM gf.guild_cfg WHERE gid = @gid LIMIT 1;";
+                cmd.Parameters.AddWithValue("gid", NpgsqlDbType.Bigint, gid);
+
+                var res = await cmd.ExecuteScalarAsync().ConfigureAwait(false);
+                if (res != null && !(res is DBNull))
+                    cid = (ulong)(long)res;
+            }
+
+            _sem.Release();
+            return cid;
+        }
+
+        public async Task SetGuildWelcomeChannelAsync(ulong gid, ulong cid)
+        {
+            await _sem.WaitAsync();
+
+            using (var con = new NpgsqlConnection(_connectionString))
+            using (var cmd = con.CreateCommand()) {
+                await con.OpenAsync().ConfigureAwait(false);
+
+                cmd.CommandText = "UPDATE gf.guild_cfg SET welcome_cid = @cid WHERE gid = @gid;";
+                cmd.Parameters.AddWithValue("gid", NpgsqlDbType.Bigint, gid);
+                cmd.Parameters.AddWithValue("cid", NpgsqlDbType.Bigint, cid);
+
+                await cmd.ExecuteNonQueryAsync().ConfigureAwait(false);
+            }
+
+            _sem.Release();
+        }
+
+        public async Task SetGuildLeaveChannelAsync(ulong gid, ulong cid)
+        {
+            await _sem.WaitAsync();
+
+            using (var con = new NpgsqlConnection(_connectionString))
+            using (var cmd = con.CreateCommand()) {
+                await con.OpenAsync().ConfigureAwait(false);
+
+                cmd.CommandText = "UPDATE gf.guild_cfg SET leave_cid = @cid WHERE gid = @gid;";
+                cmd.Parameters.AddWithValue("gid", NpgsqlDbType.Bigint, gid);
+                cmd.Parameters.AddWithValue("cid", NpgsqlDbType.Bigint, cid);
+
+                await cmd.ExecuteNonQueryAsync().ConfigureAwait(false);
+            }
+
+            _sem.Release();
+        }
+
+        public async Task RemoveGuildWelcomeChannelAsync(ulong gid)
+        {
+            await SetGuildLeaveChannelAsync(gid, 0).ConfigureAwait(false);
+        }
+
+        public async Task RemoveGuildLeaveChannelAsync(ulong gid)
+        {
+            await SetGuildLeaveChannelAsync(gid, 0).ConfigureAwait(false);
+        }
+    }
+}
