+++ TheGodfather/TheGodfather.csproj
@@ -289,7 +289,17 @@
     <Compile Include="Helpers\SWAT\SwatServerInfo.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Services\DatabaseService.cs" />
+    <Compile Include="Services\DatabaseService\BankDatabaseServices.cs" />
+    <Compile Include="Services\DatabaseService\DatabaseService.cs" />
+    <Compile Include="Services\DatabaseService\FilterDatabaseServices.cs" />
+    <Compile Include="Services\DatabaseService\GuildConfigDatabaseService.cs" />
+    <Compile Include="Services\DatabaseService\InsultDatabaseServices.cs" />
+    <Compile Include="Services\DatabaseService\MemeDatabaseService.cs" />
+    <Compile Include="Services\DatabaseService\PrefixDatabaseService.cs" />
+    <Compile Include="Services\DatabaseService\ReactionDatabaseService.cs" />
+    <Compile Include="Services\DatabaseService\StatsDatabaseService.cs" />
+    <Compile Include="Services\DatabaseService\StatusDatabaseService.cs" />
+    <Compile Include="Services\DatabaseService\SwatDatabaseService.cs" />
     <Compile Include="Services\GiphyService.cs" />
     <Compile Include="Services\ImgurService.cs" />
     <Compile Include="Services\JokesService.cs" />
